Renamed rds-record-push app to rds-services
This app will contain all code extensions for RDS
