--- rds-record-push/plugin.properties
@@ -1,4 +1,0 @@
-plugin.id=rds-record-push
-plugin.provider=Metaphacts
-plugin.version=1.0
-plugin.class=net.swissartresearch.rds.records.RecordPushPlugin
