+++ rds-services/plugin.properties
@@ -1,0 +1,4 @@
+plugin.id=rds-services
+plugin.provider=Metaphacts
+plugin.version=1.0
+plugin.class=net.swissartresearch.rds.services.RDSServicesPlugin
