+++ rds-services/src/main/resources/META-INF/services/com.metaphacts.plugin.extension.RestExtension
